Optimize platfroms for AOT
