+++ nativescript-angular/index.ts
@@ -1,4 +1,5 @@
 export * from "./platform";
+export * from "./platform-static";
 export * from "./router";
 export * from "./forms";
 export * from "./http";
